+++ tests/test_autonomy/test_cli/test_service/test_service_manager.py
@@ -27,7 +27,14 @@ from aea_test_autonomy.configurations import ETHEREUM_KEY_PATH_5
 from aea_test_autonomy.fixture_helpers import registries_scope_class  # noqa: F401
 from click.testing import Result
 
-from autonomy.chain.service import activate_service, register_instance
+from autonomy.chain.base import ServiceState
+from autonomy.chain.service import (
+    activate_service,
+    deploy_service,
+    register_instance,
+    terminate_service,
+    unbond_service,
+)
 
 from tests.test_autonomy.test_chain.base import (
     AGENT_ID,
@@ -89,6 +96,39 @@ class BaseServiceManagerTest(BaseChainInteractionTest):
             service_id=service_id,
             instances=[(agent_instance or make_crypto("ethereum").address)],
             agent_ids=[AGENT_ID],
+        )
+
+    def deploy_service(
+        self, service_id: int, deployment_payload: Optional[str] = None
+    ) -> None:
+        """Deploy service."""
+
+        deploy_service(
+            ledger_api=self.ledger_api,
+            crypto=self.crypto,
+            chain_type=self.chain_type,
+            service_id=service_id,
+            deployment_payload=deployment_payload,
+        )
+
+    def terminate_service(self, service_id: int) -> None:
+        """Terminate service."""
+
+        terminate_service(
+            ledger_api=self.ledger_api,
+            crypto=self.crypto,
+            chain_type=self.chain_type,
+            service_id=service_id,
+        )
+
+    def unbond_service(self, service_id: int) -> None:
+        """Unbond service."""
+
+        unbond_service(
+            ledger_api=self.ledger_api,
+            crypto=self.crypto,
+            chain_type=self.chain_type,
+            service_id=service_id,
         )
 
 
@@ -218,3 +258,141 @@ class TestServiceManager(BaseServiceManagerTest):
         # Will fail becaue the service is already deployed
         assert result.exit_code == 1, result.stdout
         assert "Service needs to be in finished registration state" in result.stderr
+
+    def test_terminate(
+        self,
+    ) -> None:
+        """Test terminate service."""
+
+        def _run_command(service_id: str) -> Result:
+            """Run command and return result"""
+            return self.run_cli(
+                commands=(
+                    "terminate",
+                    service_id,
+                    "--key",
+                    str(self.key_file),
+                )
+            )
+
+        result = _run_command("100")
+        assert result.exit_code == 1, result.output
+        assert "Service does not exist" in result.stderr
+
+        service_id = self.mint_service()
+
+        result = _run_command(str(service_id))
+        assert result.exit_code == 1, result.output
+        assert "Service not active" in result.stderr
+
+        self.activate_service(
+            service_id=service_id,
+        )
+        for _ in range(NUMBER_OF_SLOTS_PER_AGENT):
+            self.register_instances(
+                service_id=service_id,
+            )
+        self.deploy_service(service_id=service_id)
+
+        result = _run_command(str(service_id))
+
+        assert result.exit_code == 0, result.stderr
+        assert "Service terminated succesfully" in result.output
+
+        result = _run_command(str(service_id))
+
+        # Will fail becaue the service is already deployed
+        assert result.exit_code == 1, result.stdout
+        assert "Service already terminated" in result.stderr
+
+    def test_unbond(
+        self,
+    ) -> None:
+        """Test unbond service."""
+
+        def _run_command(service_id: str) -> Result:
+            """Run command and return result"""
+            return self.run_cli(
+                commands=(
+                    "unbond",
+                    service_id,
+                    "--key",
+                    str(self.key_file),
+                )
+            )
+
+        result = _run_command("100")
+        assert result.exit_code == 1, result.output
+        assert "Service does not exist" in result.stderr
+
+        service_id = self.mint_service()
+
+        self.activate_service(
+            service_id=service_id,
+        )
+        for _ in range(NUMBER_OF_SLOTS_PER_AGENT):
+            self.register_instances(
+                service_id=service_id,
+            )
+        self.deploy_service(service_id=service_id)
+
+        result = _run_command(str(service_id))
+        assert result.exit_code == 1, result.output
+        assert "Service needs to be in terminated-bonded state" in result.stderr
+
+        self.terminate_service(service_id=service_id)
+
+        result = _run_command(str(service_id))
+
+        assert result.exit_code == 0, result.stderr
+        assert "Service unbonded succesfully" in result.output
+
+        result = _run_command(str(service_id))
+
+        # Will fail becaue the service is already deployed
+        assert result.exit_code == 1, result.stdout
+        assert "Service needs to be in terminated-bonded state" in result.stderr
+
+    def test_info(
+        self,
+    ) -> None:
+        """Test service info."""
+
+        def _run_check(service_id: str, message: str) -> Result:
+            """Run command and return result"""
+            result = self.run_cli(commands=("info", service_id))
+            assert result.exit_code == 0
+            assert message in result.output
+
+        _run_check("100", message=ServiceState.NON_EXISTENT.name)
+
+        service_id = self.mint_service()
+        _run_check(
+            service_id=str(service_id), message=ServiceState.PRE_REGISTRATION.name
+        )
+
+        self.activate_service(service_id=service_id)
+        _run_check(
+            service_id=str(service_id), message=ServiceState.ACTIVE_REGISTRATION.name
+        )
+
+        for _ in range(NUMBER_OF_SLOTS_PER_AGENT):
+            self.register_instances(
+                service_id=service_id,
+            )
+        _run_check(
+            service_id=str(service_id), message=ServiceState.FINISHED_REGISTRATION.name
+        )
+
+        self.deploy_service(service_id=service_id)
+        _run_check(service_id=str(service_id), message=ServiceState.DEPLOYED.name)
+
+        self.terminate_service(service_id=service_id)
+        _run_check(
+            service_id=str(service_id), message=ServiceState.TERMINATED_BONDED.name
+        )
+
+        self.unbond_service(service_id=service_id)
+        _run_check(
+            service_id=str(service_id), message=ServiceState.PRE_REGISTRATION.name
+        )