+++ HISTORY.md
@@ -2,6 +2,10 @@
 
 # 0.10.5 (TBD)
 
+Autonomy:
+- Updates the agent runtime docker image to execute aea build when building the runtime image #1954
+- Make sure the alias gets used when fetching service packages #1947
+
 Docs:
 - Adds information on required agent components  #1967